+++ lib/src/eval/compiler/statement/variable_declaration.dart
@@ -1,4 +1,5 @@
 import 'package:analyzer/dart/ast/ast.dart';
+import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:dart_eval/src/eval/compiler/builtins.dart';
 import 'package:dart_eval/src/eval/compiler/context.dart';
 import 'package:dart_eval/src/eval/compiler/expression/expression.dart';
@@ -56,7 +57,12 @@ void compileVariableDeclarationList(
                 callingConvention: res.callingConvention));
       }
     } else {
-      ctx.setLocal(li.name.value() as String, BuiltinValue().push(ctx));
+      ctx.setLocal(
+          li.name.value() as String,
+          BuiltinValue()
+              .push(ctx)
+              .boxIfNeeded(ctx)
+              .copyWith(type: type ?? CoreTypes.dynamic.ref(ctx)));
     }
   }
 }